% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_output_id}
\alias{check_output_id}
\title{Check outputId}
\usage{
check_output_id(outputId)
}
\arguments{
\item{outputId}{\link{character}: The id of the table to be manipulated}
}
\description{
Check outputId is character and exist in shiny session
}
