% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{createDiffEditor}
\alias{createDiffEditor}
\title{Create diffEditor}
\usage{
createDiffEditor(
  editorAId,
  editorBId,
  elementId,
  toolBar = TRUE,
  session = shiny::getDefaultReactiveDomain(),
  sessionA = shiny::getDefaultReactiveDomain(),
  sessionB = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{editorAId}{\link{character}: The element id of the first editor}

\item{editorBId}{\link{character}: The element id of the second editor}

\item{elementId}{\link{character}: The element id of the exist widget to show diffEditor}

\item{toolBar}{\link{logical}: Show tool bar in diffEditor}

\item{session}{\link{environment}: The Shiny session object for the diffEditor (from the server function of the Shiny app)}

\item{sessionA}{\link{environment}: The Shiny session object for the first editor (from the server function of the Shiny app)}

\item{sessionB}{\link{environment}: The Shiny session object for the second editor (from the server function of the Shiny app)}
}
\value{
No return value, called for side effects
}
\description{
Create diffEditor for exist editors in an exist widget
}
\examples{
if(interactive()){
   shinyMonaco::createDiffEditor(editorAId = "editor1", editorBId = "editor2")
}
}
