% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{createDiffView}
\alias{createDiffView}
\title{Create diff editor}
\usage{
createDiffView(
  editorAId,
  editorBId,
  elementId,
  session = shiny::getDefaultReactiveDomain(),
  sessionA = shiny::getDefaultReactiveDomain(),
  sessionB = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{editorAId}{\link{character}: The element id of the first editor}

\item{editorBId}{\link{character}: The element id of the second editor}

\item{elementId}{\link{character}: The element id of the exist widget to show diffEditor}

\item{sessionA}{\link{environment}: The Shiny session object for the first editor (from the server function of the Shiny app)}

\item{sessionB}{\link{environment}: The Shiny session object for the second editor (from the server function of the Shiny app)}
}
\value{
No return value, called for side effects
}
\description{
Create diff editor for exist editors in an exist widget
}
\examples{
if(interactive()){
   shinyMonaco::createDiffView(editorAId = "editor1", editorBId = "editor2")
}
}
