% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{createMonacoDiffView}
\alias{createMonacoDiffView}
\title{Create monaco diff view}
\usage{
createMonacoDiffView(
  editorAId,
  editorBId,
  elementId,
  sessionA = shiny::getDefaultReactiveDomain(),
  sessionB = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{editorAId}{\link{character}: The element id of the first editor}

\item{editorBId}{\link{character}: The element id of the second editor}

\item{elementId}{\link{character}: The element id of the exist widget to show monacoDiffEditor}

\item{sessionA}{\link{environment}: The Shiny session object for the first editor (from the server function of the Shiny app)}

\item{sessionB}{\link{environment}: The Shiny session object for the second editor (from the server function of the Shiny app)}
}
\value{
No return value, called for side effects
}
\description{
Create monaco diff view for exist editors in an exist widget
}
\examples{
if(interactive()){
   shinyEditor::createMonacoDiffView(editorAId = "editor1", editorBId = "editor2")
}
}
