% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor.R
\name{editor}
\alias{editor}
\title{\if{html}{\out{
<Add Title>
}}}
\usage{
editor(
  value,
  enableSpellCheck = FALSE,
  fontFamily = "Consolas",
  fontSize = 16,
  insertSpaces = TRUE,
  language = "plaintext",
  lineNumbers = "on",
  placeholder = NULL,
  readOnly = FALSE,
  renderWhitespace = "boundary",
  rulers = list(100),
  scrollBeyondLastLine = TRUE,
  showStatusBar = TRUE,
  tabSize = 4,
  theme = "vs",
  wordWrap = "off",
  automaticLayout = TRUE,
  ...,
  width = "100\%",
  height = NULL,
  elementId = NULL
)
}
\description{
\if{html}{\out{
<Add Description>
}}
}
