% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{getCursorPosition}
\alias{getCursorPosition}
\title{Get cursor position in monacoEditor}
\usage{
getCursorPosition(outputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\value{
List of cursor position
}
\description{
Get cursor position in monacoEditor
}
\examples{
if(interactive()){
    shinyMonaco::getCursorPosition(outputId = "editor")
}
}
