% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMonacoLanguages}
\alias{getMonacoLanguages}
\title{Get all ace modes}
\usage{
getMonacoLanguages()
}
\value{
List of all languages in Monaco editor
}
\description{
Gets all of the available \code{modes} available in the installed version
of ace editor. Modes are often the programming or markup language which will
be used in the editor and determine things like syntax highlighting and
code folding.
}
