% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{getMonacoSelectedText}
\alias{getMonacoSelectedText}
\title{Get selected text in monacoEditor}
\usage{
getMonacoSelectedText(outputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\value{
Character of selected text
}
\description{
Get selected text in monacoEditor
}
\examples{
if(interactive()){
    shinyEditor::getMonacoSelectedText(outputId = "editor")
}
}
