% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{getMonacoSelectionRange}
\alias{getMonacoSelectionRange}
\title{Get selection range in monacoEditor}
\usage{
getMonacoSelectionRange(outputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\value{
List of selection range
}
\description{
Get selection range in monacoEditor
}
\examples{
if(interactive()){
    shinyEditor::getMonacoSelectionRange(outputId = "editor")
}
}
