% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{getSelectedText}
\alias{getSelectedText}
\title{Get selected text in editor}
\usage{
getSelectedText(outputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\value{
Character of selected text
}
\description{
Get selected text in editor
}
\examples{
if(interactive()){
    shinyMonaco::getSelectedText(outputId = "editor")
}
}
