% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getThemes}
\alias{getThemes}
\title{Get all monaco themes}
\usage{
getThemes()
}
\value{
List of all themes in Monaco editor
}
\description{
Gets all of the available \code{themes} available in the installed version
of monaco editor. Themes determine the styling and colors used in the editor.
}
