% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{onEditorReady}
\alias{onEditorReady}
\title{Fires upon editor ready}
\usage{
onEditorReady(outputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\value{
\code{TRUE}
}
\description{
Fires upon editor ready
}
\examples{
if(interactive()){
    shinyMonaco::onEditorReady(outputId = "editor")
}
}
