% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{onEditorReload}
\alias{onEditorReload}
\title{Fires upon editor reload request}
\usage{
onEditorReload(outputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\value{
\code{TRUE}
}
\description{
Fires upon editor reload request
}
\examples{
if(interactive()){
    shinyMonaco::onEditorReload(outputId = "editor")
}
}
