% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{onEditorSave}
\alias{onEditorSave}
\title{Fires upon editor save text request}
\usage{
onEditorSave(outputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\value{
\code{TRUE}
}
\description{
Fires upon editor save text request
}
\examples{
if(interactive()){
    shinyMonaco::onEditorSave(outputId = "editor")
}
}
