% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{removeDiffEditor}
\alias{removeDiffEditor}
\title{Remove diffEditor}
\usage{
removeDiffEditor(elementId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{elementId}{\link{character}: The element id of the diffEditor}

\item{session}{\link{environment}: The Shiny session object for the diffEditor (from the server function of the Shiny app)}
}
\value{
No return value, called for side effects
}
\description{
Remove monaco diff view for exist editors
}
\examples{
if(interactive()){
   shinyMonaco::removeDiffEditor(elementId = "editor1")
}
}
