% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{removeMonacoDiffView}
\alias{removeMonacoDiffView}
\title{Remove monaco diff view}
\usage{
removeMonacoDiffView(elementId)
}
\arguments{
\item{elementId}{\link{character}: The element id of the monacoDiffEditor}
}
\value{
No return value, called for side effects
}
\description{
Remove monaco diff view for exist editors
}
\examples{
if(interactive()){
   shinyMonaco::removeMonacoDiffView(elementId = "editor1")
}
}
