% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{setFocus}
\alias{setFocus}
\title{Set focus}
\usage{
setFocus(outputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\value{
No return value, called for side effects
}
\description{
Set focus to editor
}
\examples{
if(interactive()){
    shinyMonaco::setFocus(outputId = "editor")
}
}
