% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{setLanguage}
\alias{setLanguage}
\title{Set language}
\usage{
setLanguage(outputId, language, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{language}{\link{character}: The highlight of code}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\value{
No return value, called for side effects
}
\description{
Set language to editor
}
\examples{
if(interactive()){
    shinyMonaco::setLanguage(outputId = "editor", language = "text")
}
}
