% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{setMonacoLanguage}
\alias{setMonacoLanguage}
\title{Set language}
\usage{
setMonacoLanguage(
  outputId,
  language,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{language}{\link{character}: The highlight of code}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\value{
No return value, called for side effects
}
\description{
Set language to monaco editor
}
\examples{
if(interactive()){
    shinyEditor::setMonacoLanguage(outputId = "editor", language = "text")
}
}
