% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{setMonacoValue}
\alias{setMonacoValue}
\title{Replace text with new text}
\usage{
setMonacoValue(
  outputId,
  value,
  clearChangedHistory = FALSE,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{value}{\link{character}: The text of the editor}

\item{clearChangedHistory}{\link{logical}: Clear undo/redo history}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\value{
No return value, called for side effects
}
\description{
Replace all the lines in the current Document with the value of text.
}
\examples{
if(interactive()){
    shinyEditor::setMonacoValue(outputId = "editor", value = "text")
}
}
