% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{setTheme}
\alias{setTheme}
\title{Set new theme}
\usage{
setTheme(theme)
}
\arguments{
\item{theme}{\link{character}: The theme of the editor}
}
\value{
No return value, called for side effects
}
\description{
Set a new theme for the editor. theme should exist, like \code{vs-dark}
}
\examples{
if(interactive()){
    shinyMonaco::setTheme(theme = "vs")
}
}
