% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{updateMonacoOptions}
\alias{updateMonacoOptions}
\title{Update options to monacoEditor}
\usage{
updateMonacoOptions(
  outputId,
  options,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{options}{\link{list}: monaco editor options. Refer to \url{https://microsoft.github.io/monaco-editor/docs.html#interfaces/editor_editor_api.editor.IEditorOptions.html}}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\value{
No return value, called for side effects
}
\description{
Update options to monacoEditor
}
\examples{
if(interactive()){
    shinyEditor::updateMonacoOptions(outputId = "editor", options = list())
}
}
