% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{updateOptions}
\alias{updateOptions}
\title{Update options to editor}
\usage{
updateOptions(outputId, options, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{options}{\link{list}: editor options. Refer to \url{https://microsoft.github.io/monaco-editor/docs.html#interfaces/editor_editor_api.editor.IEditorOptions.html}}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\value{
No return value, called for side effects
}
\description{
Update options to editor
}
\examples{
if(interactive()){
    shinyMonaco::updateOptions(outputId = "editor", options = list())
}
}
