% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{getValue}
\alias{getValue}
\title{Get value in monacoEditor}
\usage{
getValue(outputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\value{
Character of editor text
}
\description{
Get value in monacoEditor
}
\examples{
if(interactive()){
    shinyMonaco::getValue(outputId = "editor")
}
}
